2022.11-1.EnergyUNiLAB/iEnergyUNiLAB-main/03.Sunshine/code/requirements.txt
# python 3.7.4
# https://www.python.org/ftp/python/3.7.4/python-3.7.4-macosx10.9.pkg

pandas==1.1.5
numpy==1.19.5
# https://files.pythonhosted.org/packages/a6/b7/c0594c698c7149bfe738724ab9ab3722dca3a4a43823468fe9481abe4016/numpy-1.19.5-cp38-cp38-macosx_10_9_x86_64.whl
# ERROR: numpy-1.19.5-cp38-cp38-macosx_10_9_x86_64.whl is not a supported wheel on this platform.
# https://pypi.org/project/numpy/1.19.5/#files

scikit-learn==1.0.2
paddlets==0.2.0
paddlepaddle==2.3.2
# xgboost==1.6.2
